Create botsv2
Second version of my bot in progress, it currently retweets about a key hashtag (no manual input yet). Also uploads images with text and tweets out quotes all in a x minute succession. Just need to fix the quote code for it to work properly, but this is the barebones of it all.